CLOUD-1755: Compilation fixes
<==== Mercurial data ====>
Branch: x_CLOUD-1755_update_user_auth

+++ nx_cloud/libcloud_db/src/nx/cloud/cdb/dao/memory/dao_memory_user_authentication.cpp
@@ -67,7 +67,7 @@ std::vector<std::string> UserAuthentication::fetchSystemsWithExpiredAuthRecords(
             break;
         }
 
-        if (systems.size() >= systemCountLimit)
+        if (systems.size() >= (unsigned int)systemCountLimit)
             break;
     }
